better translation check script
